Added option to gracefully shut down CAPIO server with SIGUSR1 (#157)
This pull request introduces a new mechanism to gracefully shut down the capio_server component. To use this feature, send the SIGUSR1 signal to the CAPIO server component instead of the usual SIGTERM. The capio_server instance will then:
- Reject all new handshake requests, forcing the workflow step that has sent it to terminate
- Empty the request queues so that all non blocking requests can be processed
- And finally terminate the execution.

Other changes carried out includes:
- moved from exit(EXIT_FAILURE) to using native c++ exceptions.
- Fixed compilation warnings
- Removed useless compilation targets

---------

+++ src/posix/handlers/posix_readdir.hpp
@@ -10,8 +10,8 @@
 #include <utils/requests.hpp>
 
 // Map &DIR -> <dir_path, files already served>
-inline std::unordered_map<unsigned long int, std::pair<std::string, int>> *opened_directory =
-    nullptr;
+inline std::unordered_map<unsigned long int, std::pair<std::string, long unsigned int>>
+    *opened_directory = nullptr;
 
 inline std::unordered_map<std::string, std::vector<dirent64 *> *> *directory_items;
 
@@ -50,8 +50,9 @@ inline void init_posix_dirent() {
         real_readdir = (dirent * (*) (DIR *) ) dlsym(RTLD_NEXT, "readdir");
     }
 
-    directory_items  = new std::unordered_map<std::string, std::vector<dirent64 *> *>();
-    opened_directory = new std::unordered_map<unsigned long, std::pair<std::string, int>>();
+    directory_items = new std::unordered_map<std::string, std::vector<dirent64 *> *>();
+    opened_directory =
+        new std::unordered_map<unsigned long, std::pair<std::string, unsigned long int>>();
 
     dirent_curr_dir   = new dirent64();
     dirent_parent_dir = new dirent64();
@@ -65,14 +66,14 @@ inline void init_posix_dirent() {
     syscall_no_intercept_flag = false;
 }
 
-inline int load_files_from_directory(const char *path) {
+inline unsigned long int load_files_from_directory(const char *path) {
 
     START_LOG(capio_syscall(SYS_gettid), "call(path=%s)", path);
 
     syscall_no_intercept_flag = true;
     dirent64 *entry;
-    DIR *dir  = real_opendir(path);
-    int count = 0;
+    DIR *dir                = real_opendir(path);
+    unsigned long int count = 0;
 
     if (directory_items->find(path) == directory_items->end()) {
         LOG("Directory vector not present. Adding it at path %s", path);
